DisplayTheMap,marker,popup(notInYii2)
At the beginning, try to show the map in the view in Yii2, but doesn't work.
So, I install a extension, 2amigos/yii2-google-maps-library, showing the map sucess.
but I found there's no way to get the geolocation of the users. Waste so much time to do research, try lots of approch:
1. Try to Using registerJs in Yii to run the js code from Google map api tutroial, should be runable, but can't add  "async defer" in <script> tag
2. using <script> tag in main.php to run the script, doesn't work
3. using asset bundles to run the external js code, still don't work.

So, I create another file to test the map and implement the marker,etc for test.
Will be implement thoes later in Yii2